Add files via upload
This is a program I made while inspired by the so called 'Bible Code' where hidden messages are found in the bible (or any other text). You can read more about it here: https://cutt.ly/WrtMr9R.
This program only works with Windows but since the idea isn't mine you can change the code in whatever way it needs to be changed in order to work on Linux or Mac (or any other OS). Plus, if you improve it and make it run better than I could PLEASE point me to the improved code and tell me how you improved it, I'd love to know!
I have included The King James Bible, Moby Dick and Ulysses in text file form for you to try out the program. 
Have lots of fun:-)