Add files via upload
This is a program that write coded messages. You input a 26Char string (it could be a sentence) and it uses it to code your message. The message is saved in a text file (which you name) that you can do whatever with and can be deciphered again by anybody with the pass string....or a very clever person:-) It's a bit of fun. If you wanna edit it and improve it feel free but let me know how you did that and point me to the new code:-) Cheers!